+++ app/Models/Account/Activity.php
@@ -6,7 +6,6 @@ use Parsedown;
 use App\Helpers\DateHelper;
 use App\Traits\Journalable;
 use App\Models\Contact\Contact;
-use App\Models\Journal\JournalEntry;
 use App\Interfaces\IsJournalableInterface;
 use App\Models\ModelBindingHasher as Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -78,14 +77,6 @@ class Activity extends Model implements IsJournalableInterface
     public function type()
     {
         return $this->belongsTo(ActivityType::class, 'activity_type_id');
-    }
-
-    /**
-     * Get all of the activities journal entries.
-     */
-    public function journalEntries()
-    {
-        return $this->morphMany(JournalEntry::class, 'journalable');
     }
 
     /**