+++ tests/Unit/Models/EntryTest.php
@@ -12,12 +12,12 @@ class EntryTest extends TestCase
 
     public function test_get_info_for_journal_entry()
     {
-        $entry = factory(Entry::class)->make();
-        $entry->id = 1;
-        $entry->title = 'This is the title';
-        $entry->post = 'this is a post';
-        $entry->created_at = '2017-01-01 00:00:00';
-        $entry->save();
+        $entry = factory(Entry::class)->make([
+            'id' => 1,
+            'title' => 'This is the title',
+            'post' => 'this is a post',
+            'created_at' => '2017-01-01 00:00:00',
+        ]);
 
         $data = [
             'type' => 'entry',
@@ -29,6 +29,7 @@ class EntryTest extends TestCase
             'month' => 1,
             'month_name' => 'JAN',
             'year' => 2017,
+            'date' => '2017-01-01 00:00:00',
         ];
 
         $this->assertEquals(